--- examples/yunba-1.0.1.js
@@ -1,1 +1,0 @@
-../yunba-1.0.1.js
